stop vm and relevant qemu cleanup
