+++ src/backend/mod.rs
@@ -11,10 +11,7 @@ pub mod qemu;
 pub mod qemu_old;
 pub mod tart;
 
-use crate::{
-    global_paths::VciGlobalPaths,
-    vm_image::{Arch, GuestOs, SshTarget},
-};
+use crate::vm_image::{Arch, GuestOs, SshTarget};
 
 #[derive(Debug, Clone, Default)]
 pub struct VmStartConfig {
