+++ src/backend/qemu/backend.rs
@@ -37,21 +37,21 @@ pub struct QemuBackend {
     /// and routed to stdio when `!graphics`, or to this file when `graphics`.
     pub serial_log: Option<TargetPath>,
     pub exec_target: HostExecTarget,
-    host_temp_dir: PathBuf,
-    exec_target_temp_dir: TargetPath,
+    pub host_temp_dir: PathBuf,
+    pub exec_target_temp_dir: TargetPath,
 
     /// Shared or exclusive `vci_image_<name>.lock`
-    image_lock: FileLock,
-    host_port: Option<PortFlock>,
-    disk: BackingFile,
-    uefi_vars: Option<BackingFile>,
-    additional_drives: Vec<AdditionalDrive>,
-    tpm_info: Option<TpmInfo>,
+    pub image_lock: FileLock,
+    pub host_port: Option<PortFlock>,
+    pub disk: BackingFile,
+    pub uefi_vars: Option<BackingFile>,
+    pub additional_drives: Vec<AdditionalDrive>,
+    pub tpm_info: Option<TpmInfo>,
 
-    orphans: OrphanTracker,
-    qemu_process: Option<Arc<Mutex<TargetChildProcess>>>,
+    pub orphans: OrphanTracker,
+    pub qemu_process: Option<Arc<Mutex<TargetChildProcess>>>,
     /// Must be stored here, as it must restart when QEMU process restarts.
-    tpm_process: Option<Arc<Mutex<TargetChildProcess>>>,
+    pub tpm_process: Option<Arc<Mutex<TargetChildProcess>>>,
 }
 
 impl QemuBackend {
@@ -250,7 +250,7 @@ impl VmBackend for QemuBackend {
             .join(format!("{}-qemu.stderr", self.run_marker()));
 
         let qemu = loop {
-            let cmd = build_qemu_args(self)?;
+            let cmd = super::binaries::build_qemu_args(self)?;
             let qemu = TargetChildProcess::new(
                 &self.exec_target,
                 &self.run_marker(),
@@ -261,9 +261,9 @@ impl VmBackend for QemuBackend {
             .context("Failed to spawn QEMU")?;
             self.orphans.add_child_process(&qemu);
 
-            match qemu_launch_outcome(&qemu, &stderr_log)? {
-                QemuLaunchOutcome::Running => break qemu,
-                QemuLaunchOutcome::PortTaken => {
+            match super::binaries::qemu_launch_outcome(&qemu, &stderr_log)? {
+                super::binaries::QemuLaunchOutcome::Running => break qemu,
+                super::binaries::QemuLaunchOutcome::PortTaken => {
                     // QEMU already exited, so drop it, release the reserved port, and try again.
                     drop(qemu);
                     let taken = self.host_port.as_ref().expect("port reserved").port;
@@ -280,7 +280,26 @@ impl VmBackend for QemuBackend {
     }
 
     fn stop_vm(&mut self) {
-        todo!()
+        if let Some(qemu) = self.qemu_process.take() {
+            if let Ok(mut guard) = qemu.lock() {
+                guard.kill();
+            }
+        }
+        if let Some(tpm) = self.tpm_process.take() {
+            if let Ok(mut guard) = tpm.lock() {
+                guard.kill();
+            }
+        }
+
+        // Directly remove swtpm's control socket, cause SIGKILL may not unlink it.
+        // This allows a restarted swtpm to be able to recreate it.
+        if let Some(info) = &self.tpm_info {
+            let _ = super::binaries::target_command(&self.exec_target, "rm")
+                .args(["-f", &info.socket_path.native_path()])
+                .status();
+        }
+
+        self.host_port = None;
     }
 
     fn ssh_target(&self) -> crate::vm_image::SshTarget {
@@ -304,7 +323,33 @@ impl VmBackend for QemuBackend {
     }
 
     fn wait_for_exit(&mut self) {
-        todo!()
+        if let Some(qemu) = self.qemu_process.clone() {
+            loop {
+                let exited = qemu
+                    .lock()
+                    .map(|mut guard| guard.try_wait())
+                    .unwrap_or(true);
+                if exited {
+                    break;
+                }
+                std::thread::sleep(std::time::Duration::from_millis(200));
+            }
+        }
+
+        self.qemu_process = None;
+        if let Some(tpm) = self.tpm_process.take() {
+            if let Ok(mut guard) = tpm.lock() {
+                guard.kill();
+            }
+        }
+
+        // Directly remove swtpm's control socket, cause SIGKILL may not unlink it.
+        // This allows a restarted swtpm to be able to recreate it.
+        if let Some(info) = &self.tpm_info {
+            let _ = super::binaries::target_command(&self.exec_target, "rm")
+                .args(["-f", &info.socket_path.native_path()])
+                .status();
+        }
     }
 
     fn serial_log_path(&self) -> Option<&Path> {
@@ -315,6 +360,50 @@ impl VmBackend for QemuBackend {
     }
 }
 
+impl Drop for QemuBackend {
+    fn drop(&mut self) {
+        if let Some(qemu) = self.qemu_process.take() {
+            if let Ok(mut guard) = qemu.lock() {
+                guard.kill();
+            }
+        }
+        if let Some(tpm) = self.tpm_process.take() {
+            if let Ok(mut guard) = tpm.lock() {
+                guard.kill();
+            }
+        }
+
+        // [`BackingFile::temp()`] returns a path only for `Temp`.
+        let mut files: Vec<PathBuf> = Vec::new();
+        if let Some(p) = self.disk.temp() {
+            files.push(p.path.clone());
+        }
+        if let Some(p) = self.uefi_vars.as_ref().and_then(|v| v.temp()) {
+            files.push(p.path.clone());
+        }
+        for drive in &self.additional_drives {
+            if let Some(p) = drive.file.temp() {
+                files.push(p.path.clone());
+            }
+        }
+        if let Some(serial) = &self.serial_log {
+            files.push(serial.path.clone());
+        }
+        files.push(
+            self.host_temp_dir
+                .join(format!("{}-qemu.stderr", self.run_marker())),
+        );
+
+        for path in files {
+            let _ = std::fs::remove_file(&path);
+        }
+
+        if let Some(info) = &self.tpm_info {
+            let _ = std::fs::remove_dir_all(&info.state_dir.path);
+        }
+    }
+}
+
 /// A file the VM boots from, resolved to the `exec_target`'s namespace fully. The variant is the
 /// source of truth for whether cleanup should delete it or not.
 /// - [`BackingFile::Base`] is the image's own file, used in place so writes persist and it is
@@ -322,21 +411,21 @@ impl VmBackend for QemuBackend {
 /// - [`BackingFile::Temp`] is a per-run throwaway in the temp dir (a qcow2 overlay for disks, a
 ///   plain copy for UEFI vars) so writes are discarded and it is deleted on cleanup.
 ///   Always for `virtci run`.
-enum BackingFile {
+pub enum BackingFile {
     Base(TargetPath),
     Temp(TargetPath),
 }
 
 impl BackingFile {
     /// The path QEMU should use, in either mode.
-    fn target(&self) -> &TargetPath {
+    pub fn target(&self) -> &TargetPath {
         match self {
             BackingFile::Base(p) | BackingFile::Temp(p) => p,
         }
     }
 
     /// `Some` only for a throwaway file that cleanup is responsible for deleting.
-    fn temp(&self) -> Option<&TargetPath> {
+    pub fn temp(&self) -> Option<&TargetPath> {
         match self {
             BackingFile::Temp(p) => Some(p),
             BackingFile::Base(_) => None,
@@ -346,15 +435,15 @@ impl BackingFile {
 
 /// One extra `-drive`, such as the macOS OpenCore bootloader. `spec` is the full `-drive` arg
 /// with its `file=` already pointing at `file`'s in-namespace path.
-struct AdditionalDrive {
-    spec: String,
-    file: BackingFile,
+pub struct AdditionalDrive {
+    pub spec: String,
+    pub file: BackingFile,
 }
 
-struct TpmInfo {
-    state_dir: TargetPath,
+pub struct TpmInfo {
+    pub state_dir: TargetPath,
     /// swtpm's control socket inside of `state_dir`. swtpm and QEMU share it.
-    socket_path: TargetPath,
+    pub socket_path: TargetPath,
 }
 
 struct RunSetup {
@@ -658,263 +747,4 @@ fn temp_dir_target(paths: &VciGlobalPaths, exec_target: &HostExecTarget) -> Targ
             path: paths.temp.clone(),
         }
     }
-}
-
-struct QemuBuiltCommand {
-    program: String,
-    arguments: Vec<String>,
-}
-
-/// Append a QEMU `-flag value` pair. MUST be two separate argv strings.
-fn push_arg(args: &mut Vec<String>, flag: &str, value: impl Into<String>) {
-    args.push(flag.to_string());
-    args.push(value.into());
-}
-
-/// Build the full `qemu-system-*` invocation for `backend`, ready to spawn on its `exec_target`.
-/// Requires the SSH port to be acquired (`host_port`) and cpus/memory/offline resolved into
-/// `start_config`.
-fn build_qemu_args(backend: &QemuBackend) -> anyhow::Result<QemuBuiltCommand> {
-    use super::{binaries, kvm};
-    use crate::vm_image::{Arch, GuestOs};
-
-    let qemu_config = backend
-        .base_image
-        .backend
-        .as_qemu()
-        .expect("Expected QEMU config");
-    let arch = backend.base_image.arch;
-
-    let (program, _version) =
-        binaries::qemu_system_binary(arch, &backend.exec_target, backend.is_tpm())?;
-
-    let mut args = Vec::<String>::new();
-
-    push_arg(
-        &mut args,
-        "-machine",
-        binaries::qemu_machine(arch, &backend.exec_target),
-    );
-
-    let cpu = match &qemu_config.cpu_model {
-        Some(model) => model.clone(),
-        None => binaries::qemu_cpu(arch, &backend.exec_target).to_string(),
-    };
-    push_arg(&mut args, "-cpu", cpu);
-
-    push_arg(&mut args, "-name", backend.run_marker());
-
-    let cpus = backend
-        .start_config
-        .cpus
-        .expect("cpus resolved before launch");
-    let memory_mb = backend
-        .start_config
-        .memory_mb
-        .expect("memory resolved before launch");
-    push_arg(&mut args, "-smp", cpus.to_string());
-    push_arg(&mut args, "-m", format!("{memory_mb}M"));
-
-    if let Some(uefi) = &qemu_config.uefi {
-        push_arg(
-            &mut args,
-            "-drive",
-            format!("if=pflash,format=raw,unit=0,readonly=on,file={}", uefi.code),
-        );
-        if let Some(vars) = &backend.uefi_vars {
-            push_arg(
-                &mut args,
-                "-drive",
-                format!(
-                    "if=pflash,format=raw,unit=1,file={}",
-                    vars.target().native_path()
-                ),
-            );
-        }
-    }
-
-    // Extra drives like OpenCore. The spec already has `file=` in the exec namespace.
-    for drive in &backend.additional_drives {
-        push_arg(&mut args, "-drive", drive.spec.clone());
-    }
-
-    let disk = backend.disk.target().native_path();
-    if qemu_config.additional_devices.is_some() {
-        // Config is supplying it's own `-device` list.
-        // TODO investigate correctness of this with many drives/devices
-        push_arg(
-            &mut args,
-            "-drive",
-            format!("id=SystemDisk,if=none,file={disk},format=qcow2"),
-        );
-    } else {
-        match arch {
-            Arch::ARM64 | Arch::RISCV64 => {
-                push_arg(
-                    &mut args,
-                    "-drive",
-                    format!("id=SystemDisk,if=none,file={disk},format=qcow2"),
-                );
-                if qemu_config.nvme {
-                    push_arg(
-                        &mut args,
-                        "-device",
-                        "nvme,drive=SystemDisk,serial=SystemDisk,bootindex=0",
-                    );
-                } else {
-                    push_arg(
-                        &mut args,
-                        "-device",
-                        "virtio-blk-pci,drive=SystemDisk,bootindex=0",
-                    );
-                }
-            }
-            Arch::X64 => {
-                if qemu_config.nvme {
-                    push_arg(
-                        &mut args,
-                        "-drive",
-                        format!("id=SystemDisk,if=none,file={disk},format=qcow2"),
-                    );
-                    push_arg(
-                        &mut args,
-                        "-device",
-                        "nvme,drive=SystemDisk,serial=SystemDisk,bootindex=0",
-                    );
-                } else if backend.base_image.os == GuestOs::Windows {
-                    push_arg(
-                        &mut args,
-                        "-drive",
-                        format!("file={disk},format=qcow2,if=ide"),
-                    );
-                } else {
-                    push_arg(
-                        &mut args,
-                        "-drive",
-                        format!("file={disk},format=qcow2,if=virtio"),
-                    );
-                }
-            }
-        }
-    }
-
-    if let Some(isos) = &qemu_config.readonly_isos {
-        for iso in isos {
-            push_arg(
-                &mut args,
-                "-drive",
-                format!("file={iso},format=raw,if=virtio,readonly=on"),
-            );
-        }
-    }
-
-    if !backend.graphics {
-        push_arg(&mut args, "-display", "none");
-    }
-
-    if let Some(serial_log) = &backend.serial_log {
-        if backend.graphics {
-            push_arg(
-                &mut args,
-                "-serial",
-                format!("file:{}", serial_log.native_path()),
-            );
-        } else {
-            push_arg(&mut args, "-serial", "stdio");
-        }
-    }
-
-    push_arg(&mut args, "-rtc", "base=utc");
-
-    match backend.exec_target {
-        HostExecTarget::Linux | HostExecTarget::WSL2(_) => {
-            if kvm::check_kvm_access(&backend.exec_target).is_ok() {
-                push_arg(&mut args, "-accel", "kvm");
-            }
-        }
-        HostExecTarget::MacOS => push_arg(&mut args, "-accel", "hvf"),
-        HostExecTarget::WindowsNative => push_arg(&mut args, "-accel", "whpx"),
-    }
-    push_arg(&mut args, "-accel", "tcg");
-
-    let host_port = backend
-        .host_port
-        .as_ref()
-        .expect("host port acquired before build_qemu_args")
-        .port;
-    let inside = backend.inside_vm_port;
-    let offline = backend.start_config.offline.unwrap_or(false);
-    let netdev = if offline {
-        format!("user,id=net0,restrict=yes,hostfwd=tcp::{host_port}-:{inside}")
-    } else {
-        format!("user,id=net0,hostfwd=tcp::{host_port}-:{inside}")
-    };
-    push_arg(&mut args, "-netdev", netdev);
-
-    // `disable-modern=on` forces virtio legacy (INTx, not MSI-X), required by WHPX.
-    if matches!(backend.exec_target, HostExecTarget::WindowsNative) {
-        push_arg(
-            &mut args,
-            "-device",
-            "virtio-net-pci,netdev=net0,disable-modern=on",
-        );
-    } else {
-        push_arg(&mut args, "-device", "virtio-net-pci,netdev=net0");
-    }
-
-    if let Some(tpm) = &backend.tpm_info {
-        push_arg(
-            &mut args,
-            "-chardev",
-            format!("socket,id=chrtpm,path={}", tpm.socket_path.native_path()),
-        );
-        push_arg(&mut args, "-tpmdev", "emulator,id=tpm0,chardev=chrtpm");
-        let device = match arch {
-            Arch::ARM64 | Arch::RISCV64 => "tpm-tis-device,tpmdev=tpm0",
-            Arch::X64 => "tpm-tis,tpmdev=tpm0",
-        };
-        push_arg(&mut args, "-device", device);
-    }
-
-    if let Some(devices) = &qemu_config.additional_devices {
-        for device in devices {
-            push_arg(&mut args, "-device", device.clone());
-        }
-    }
-
-    Ok(QemuBuiltCommand {
-        program,
-        arguments: args,
-    })
-}
-
-enum QemuLaunchOutcome {
-    Running,
-    PortTaken,
-}
-
-/// Observe the launch for about 1.5 seconds to see if the port was taken.
-fn qemu_launch_outcome(
-    qemu: &Arc<Mutex<TargetChildProcess>>,
-    stderr_log: &Path,
-) -> anyhow::Result<QemuLaunchOutcome> {
-    const POLL: std::time::Duration = std::time::Duration::from_millis(100);
-    const ATTEMPTS: u32 = 15;
-
-    for _ in 0..ATTEMPTS {
-        std::thread::sleep(POLL);
-        let exited = {
-            let mut guard = qemu.lock().expect("QEMU process lock poisoned how?");
-            guard.try_wait()
-        };
-        if exited {
-            let stderr = std::fs::read_to_string(stderr_log).unwrap_or_default();
-            if stderr.to_lowercase().contains("host forwarding rule") {
-                return Ok(QemuLaunchOutcome::PortTaken);
-            }
-            anyhow::bail!("QEMU exited immediately after launch:\n{}", stderr.trim());
-        }
-    }
-
-    Ok(QemuLaunchOutcome::Running)
 }
